Bug Fix: ReadInput() only didn't throw error if user quit
This caused some errors to be handled twice.

Another change made was handling the error of an invalid session ID
in the client mode. Before this commit, it would kill the program
but now it will just loop back and try again

Another change made was "expiring" error messages after 3 seconds.
This means the error will not persist onscreen afterwards, which is
just a QOL thing.