Fix NO_COLOR logic
The official documentation for NO_COLOR says that it requires a
non-empty NO_COLOR environment variable, but I was just checking
to see if such a environment variable existed. Now that I am checking
for the contents, however, I need not use EnvLookup, and I can just
use Getenv instead.