themes!
Blur now uses configuration files. For now, all they specify is
a theme, but in the future, they should be able to help out more.

I still have to write some code that checks if the blur directory
exists and if not, creates it and puts all the themes in it.